restructure utils
